requirements.txt
httpx>=0.27.0
rich>=13.0.0
python-dotenv>=1.0.0
dnspython>=2.6.0
fastapi>=0.111.0
uvicorn>=0.29.0
